+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ### mTroll MIDI controller changelog / process milestones
 
+#### 2026.xx.xx
+- Trace window now displays MIDI device add/remove change status messages
+- Updated WinMM MIDI error reporting
+
 #### 2026.03.04
 - Added support for importing patches from another file so that, for example, device patches don't need to be copied/pasted across configs [#60](../../../issues/60)
 - If multiple patches are identically named, the first one wins (gets loaded), rather than the last